+++ index.html
@@ -158,20 +158,30 @@
     </div>
     
 
+    <!-- Basic Dependencies -->
     <script src="js/vendor/jquery-2.1.0.min.js"></script>
     <script src="js/vendor/underscore.js"></script>
     <script src="js/vendor/backbone-1.1.0.js"></script>
     <script src="js/vendor/bootstrap-tooltip.min.js"></script>
+
+    <!-- Settings Module and API Keys -->
     <script src="js/vendor/settings.js"></script>
     <script src="js/vendor/config.js"></script>
-    <script src="js/app.js"></script>
 
-    <script src="js/vendor/video-js/video.dev.js"></script>
-    <script src="js/frontend/app.js"></script>
+    <!-- App Initialization -->
+    <script src="js/app.js"></script>
+    <script src="js/language.js"></script>
+    <script src="js/tracking.js"></script>
+    <script src="js/updater.js"></script>
+
+    <!-- UI and UI-related Initialization -->
+    <script src="js/frontend/ui.js"></script>
+    <script src="js/frontend/player.js"></script>
+
+    <!-- Data Sources -->
     <script src="js/frontend/providers/torrents.js"></script>
-    <script src="js/frontend/providers/cache.js"></script>
-    <script src="js/frontend/providers/themoviedb.js"></script>
-    <script src="js/frontend/providers/yifysubtitles.js"></script>
+
+    <!-- Backbone Models, Views and Controllers -->
     <script src="js/frontend/models/movie.js"></script>
     <script src="js/frontend/views/page.js"></script>
     <script src="js/frontend/views/listview.js"></script>
@@ -181,7 +191,14 @@
     <script src="js/frontend/controllers/search.js"></script>
     <script src="js/frontend/controllers/filter.js"></script>
     <script src="js/frontend/router.js"></script>
+
+    <!-- Remote Control -->
     <script src="rc/index.js"></script>
+
+    <!-- Other Dependencies -->
+    <script src="js/vendor/video-js/video.dev.js"></script>
+    <script src="js/vendor/dragwindow.js"></script>
+
     <script>userTracking.event('App Start', 'Ready - '+Settings.get('version'),  (Math.round(((new Date()*1) - __startTime) / 100) * 100) +'ms' ).send();</script>
   </body>
 </html>
